More UA client cleanup.
